@@ -1,13 +1,12 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import React from 'react'
+import { createRoot } from 'react-dom'
 
-import './index.css';
-import App from './App';
-import { ContextProvider } from './contexts/ContextProvider';
+import './index.css'
+import App from './App'
+import { ContextProvider } from './contexts/ContextProvider'
 
-ReactDOM.render(
-	<ContextProvider>
-		<App />
-	</ContextProvider>,
-	document.getElementById('root')
-);
+createRoot(document.getElementById('root')).render(
+  <ContextProvider>
+    <App />
+  </ContextProvider>
+)
